1. splitup the demos.py into individual ones. 2. changed the way of demo handling, which is more optimized for extensibility. 3. fix the svr demo bug.
